backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
firebase-admin
google-generativeai
python-multipart
google-cloud-storage
Pillow
